Prevent re-listing in less than 30 days.
The words in a listing are now scraped into a bag of words. All new listings are now checked against previous listings from the same email domain. If there is a more than 60% match in the bags of words, the listing is refused. We use this approach instead of a direct comparison to make it more robust. Listings will not be considered unique with simple formatting changes, by shifting words around, or by substituting a few throwaway lines. They'll need to be substantially different. Listings are checked against the email domain rather than email address because new addresses are cheap while new domains are not.

This commit also fixes a long known bug that allowed the email address to be changed after a listing was published.